list schedule refactor + move loadingdatadisplay back to listschedule
